query as fetch instead of buildtime
This will reduce this report's cost to total build time for the site.

Additionally, I've abstraced the SearchField into a re-usable component, to eventually use in the other reports.
